feat(monocle-backend,monocle-nextjs): Export MonocleAssessment type
