+++ .changeset/angry-clouds-show.md
@@ -1,0 +1,6 @@
+---
+"@spur.us/monocle-backend": minor
+"@spur.us/monocle-nextjs": minor
+---
+
+Add `MonocleAssessment` type
