+++ packages/monocle-backend/src/types.ts
@@ -1,0 +1,9 @@
+/**
+ * Configuration options for creating a MonocleClient instance
+ */
+export interface MonocleOptions {
+  /** The secret key used for authentication with the Monocle API */
+  secretKey: string;
+  /** Optional base URL for the Monocle API. Defaults to the standard API URL if not provided */
+  baseUrl?: string;
+}
